# programmazione-web-lab/WASAtext/doc/api.yaml
openapi: 3.0.3
info:
  title: WASAtext 
  version: "1.0"
  description: WASAtext application permette di scambiare messaggi e foto a tutti coloro che sono registrati nell'applicazione.
servers:
  - url: baseURL
    description: server principale di wasa text application
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: string
...
  responses:
...
    OK:
      description: Operazione completata con successo
      content:
        application/json:
          schema:
            type: string
            example: "Operazione riuscita"

    # --- 201 Created (Successo creazione) ---
...
    BadRequest:
      description: Richiesta malformata o parametri mancanti/errati
      content:
        application/json:
          schema:
            type: string
            example: "Parametri della richiesta non validi"

    # --- 401 Unauthorized (Manca autenticazione) ---
    Unauthorized:
      description: Autenticazione fallita o token mancante
      content:
        application/json:
          schema:
            type: string
            example: "Non autorizzato: effettua il login"

    # --- 403 Forbidden (Permessi mancanti) ---
    Forbidden:
      description: L'utente non ha i permessi per questa azione
      content:
        application/json:
          schema:
            type: string
            example: "Accesso negato: azione non consentita"

    # --- 404 Not Found (Risorsa inesistente) ---
    NotFound:
      description: La risorsa richiesta non è stata trovata
      content:
        application/json:
          schema:
            type: string
            example: "Risorsa non trovata"

    # --- 500 Internal Server Error (Errore DB o Crash) ---
    InternalError:
      description: Errore interno del server
      content:
        application/json:
          schema:
            type: string
            example: "Errore interno del server: riprova più tardi"
security:
  - bearerAuth: []
paths:
...
  /users/{userId}/groups/{groupId}/photo:
    parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            description: id univoco per l'user
            example: "utente123"
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            description: chat id univoco per recuperare un gruppo
            example: "group123"
    get:
      description: retrive della foto profilo del gruppo
      operationId: getGroupPhoto
      tags: ["getGroupPhoto"]
      responses:
          "200":
            description: restituisce la foto del gruppo
            content:
              application/json:
                schema:
                  type: string
                  example: "path/to/photo"
          '400':
            $ref: '#/components/responses/BadRequest'
          '401':
            $ref: '#/components/responses/Unauthorized'
          '403':
            $ref: '#/components/responses/Forbidden'
          '404':
            $ref: '#/components/responses/NotFound'
          '500':
            $ref: '#/components/responses/InternalError'
      
    put:
      description: cambia foto del gruppo
      operationId: setGroupPhoto
      tags: ["setGroupPhoto"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              example: "path/to/photo"
      responses:
        "200":
          description: foto grupppo cambiata
          content:
            application/json:
              schema:
                type: string
                example: "foto gruppo cambiata"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{userId}/groups/{groupId}/members/{targetUserId}:
    parameters:
        - name: targetUserId
          in: path
          required: true
          schema:
            type: string
            description: id univoco per l'user che si vuole aggiungere o rimuovere
            example: "utente1231"
        - name: userId
          in: path
          required: true
          schema:
            type: string
            description: id univoco per l'user
            example: "utente123"
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            description: chat id univoco per recuperare un gruppo
            example: "group123"
    delete:
      description: rimuove un partecipante dal gruppo
      operationId: removeFromGroup
      tags: ["removeFromGroup"]
      responses:
        "200":
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...